+++ src/main/java/legend/game/combat/effects/LightningBoltEffectSegmentOrigin08.java
@@ -4,8 +4,8 @@ package legend.game.combat.effects;
 public class LightningBoltEffectSegmentOrigin08 {
   public final int index;
 
-  public int x_00;
-  public int y_04;
+  public float x_00;
+  public float y_04;
 
   public LightningBoltEffectSegmentOrigin08(final int index) {
     this.index = index;
