+++ src/main/java/legend/game/combat/effects/ThunderArrowEffectBolt1e.java
@@ -3,8 +3,8 @@ package legend.game.combat.effects;
 import legend.core.gte.USCOLOUR;
 
 public class ThunderArrowEffectBolt1e {
-  public short x_00;
-  public short y_02;
+  public float x_00;
+  public float y_02;
   public final USCOLOUR colour_04 = new USCOLOUR();
   public final USCOLOUR colour_0a = new USCOLOUR();
   public final USCOLOUR svec_10 = new USCOLOUR();
